[FIX] ASAN: container overflow
We reserve a vector and have an access via [] that is within the capacity, but out of the size; e.g., this warning is triggered for a vector of capacity 10 and size 4 when for any acces in [4, 10). The vector in question is only needed once, all other callees do not need the vector.